Created the Join and Leave button to handle missions process.
